backend/requirements.txt
fastapi
langchain-google-genai
langchain
uvicorn
python-dotenv